+++ examples/components/src/__tests__/character.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { render } from 'react-testing-library';
-import Character from '../character';
+import Character from '../Character';
 
 import {
   empty,