+++ test/test-utils/enzyme.ts
@@ -1,4 +1,4 @@
-import Enzyme from 'enzyme';
+import * as Enzyme from 'enzyme';
 const Adapter = require('enzyme-adapter-react-16');
 
 Enzyme.configure({ adapter: new Adapter() });